+++ template/hello/template/bin/main.ml
@@ -1,0 +1,3 @@
+(** Main entry point for our application. *)
+
+let () = print_endline @@ Hello.greet "World"