+++ template/hello/template/lib/hello.mli
@@ -1,0 +1,8 @@
+(** A short description of the project *)
+
+val greet : string -> string
+(** Returns a greeting message.
+
+    {4 Examples}
+
+    {[ print_endline @@ greet "Jane" ]} *)