+++ template/spin_template.mli
@@ -6,5 +6,8 @@ module type Template = sig
   val read : string -> string option
 end
 
+val hello : (module Template)
+(** The hello tutorial template *)
+
 val all : (module Template) list
 (** List of all of the official spin templates *)